chore: move public/svg to src/svg
