+++ src/components/loading-indicator.ts
@@ -2,7 +2,7 @@ import { LitElement, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { styles } from '../styles/loading-indicator.js';
 import { unsafeSVG } from 'lit/directives/unsafe-svg.js';
-import iconSpinner from '../../public/svg/spinner-icon.svg?raw';
+import iconSpinner from '../svg/spinner-icon.svg?raw';
 
 @customElement('loading-indicator')
 export class LoadingIndicatorComponent extends LitElement {
