+++ src/svg/readme.md
